Better typing

+++ tests/test_policies.py
@@ -2,16 +2,15 @@ import numpy as np
 import marl
 from .utils import almost_equal
 
+
 def test_softmax():
-    p = marl.policy.SoftmaxPolicy(3, 1.)
-    available_actions = np.array([
-        [1, 1, 0],
-        [0, 0, 1]
-    ])
+    p = marl.policy.SoftmaxPolicy(3, 1.0)
+    available_actions = np.array([[1, 1, 0], [0, 0, 1]])
     for _ in range(1_000):
-        qvalues = np.random.rand(2, 3)
+        qvalues = np.random.rand(2, 3).astype(np.float32)
         action = p.get_action(qvalues, available_actions)
         assert 0 <= action[0] <= 1 and action[1] == 2
+
 
 def test_egreedy_update():
     p = marl.policy.EpsilonGreedy.linear(1, 0.1, 10)
